It is a simple Shoot'em up (SHMUP)-like space game made using Python and the pygame library.
It is a simple Shoot'em up (SHMUP)-like space game made using Python and the pygame library. In such games, the player usually controls a spaceship or plane at the bottom of the screen and fights against waves of enemies coming from above. The aim of the game in the project is to collect points by shooting enemies coming from above with the spaceship the player controls. If the player hits one of the enemy ships, the game ends.